warnings: fix signed/unsigned comparison warnings

+++ Rendering/OpenGL2/vtkOpenGLPolyDataMapper2D.cxx
@@ -614,7 +614,7 @@ void vtkOpenGLPolyDataMapper2D::RenderOverlay(vtkViewport* viewport,
     glLineWidth(actor->GetProperty()->GetLineWidth()); // supported by all OpenGL versions
 
     this->Lines.ibo.Bind();
-    for (int eCount = 0; eCount < this->Lines.offsetArray.size(); ++eCount)
+    for (size_t eCount = 0; eCount < this->Lines.offsetArray.size(); ++eCount)
       {
       glDrawElements(GL_LINE_STRIP,
         this->Lines.elementsArray[eCount],
@@ -639,7 +639,7 @@ void vtkOpenGLPolyDataMapper2D::RenderOverlay(vtkViewport* viewport,
   if (this->TriStrips.indexCount)
     {
     this->TriStrips.ibo.Bind();
-    for (int eCount = 0; eCount < this->TriStrips.offsetArray.size(); ++eCount)
+    for (size_t eCount = 0; eCount < this->TriStrips.offsetArray.size(); ++eCount)
       {
       glDrawElements(GL_TRIANGLE_STRIP,
         this->TriStrips.elementsArray[eCount],
